+++ src/common/types/constructor-type.ts
