+++ src/data-access/contracts/database-interface.ts
@@ -1,0 +1,6 @@
+import { Constructor } from "../../common/types/constructor-type";
+import { DbSet } from "../db-set";
+
+export interface IDatabase {
+    getSet<T>(type: Constructor<T>): DbSet<T>;
+}
