--- src/data-access/database-interface.ts
@@ -1,6 +1,0 @@
-import { Constructor } from "../common/constructor-type";
-import { DbSet } from "./db-set";
-
-export interface IDatabase {
-    getSet<T>(type: Constructor<T>): DbSet<T>;
-}
