+++ src/data-access/unit-of-work.ts
@@ -1,8 +1,8 @@
-import { Constructor } from "../common/constructor-type";
-import { IDatabase } from "./database-interface";
+import { Constructor } from "../common/types/constructor-type";
+import { IDatabase } from "./contracts/database-interface";
 import { Repository } from "./repository";
-import { IRepository } from "./repositoy-interface";
-import { IUnitOfWork } from "./unit-of-work-interface";
+import { IRepository } from "./contracts/repositoy-interface";
+import { IUnitOfWork } from "./contracts/unit-of-work-interface";
 
 export class UnitOfWork implements IUnitOfWork {
     public static inject = ["database"] as const;
